// src/types/projectMetadata.ts
export type ProjectMetadata = {
  title: string;
  description: string;
  website: string;
  logoImg: string;
  bannerImg: string;
  projectTwitter: string;
  userGithub: string;
  projectGithub: string;
  projectTelegram: string;
  projectWarpcast: string;
  projectGuild: string;
  projectDiscord: string;
  projectLens: string;
};